Mickey fixed the problem that template error orrured while looking up journals at back end.
Without setting up any receipt printer, the error would occur in every
processed template, while only the first one went wrong if there was a
receipt printer set.

It is because _templateModifiers( TrimPath, encoding ) will be invoked
in method printReceipts and the journal is processed before printing receipt.

Therefore, we may call _templateModifiers right before processing tpl
in journal controller. By doing so we can now acquire the proper
'encoding', that is, when processing an html tpl, the encoding will be
UTF-8, on the other hand, we read the encoding from the selected device
when processing tpl for printing.