Fixed bug #73
Bug is due to the fact that our data export script relies on sqlite3's
built-in INSERT output mode to generate SQL statements that we later
invoke during the import phase to insert data back into the database.
sqlite3, unfortunately, only examines the actual record values to decide
if it needs to put quotes around the value in the insert statement. If
the value happens to be a string representation of a number, then
sqlite3 outputs it as a number. This will cause the INSERT statement to
process the value as a number even if the field definition is VARCHAR.

The data import/export process for raw table has been updated to use
tabs as field separators.

Also, import/export of ACL permissions now exclude acl_users table.
After import of ACL permissions, it is recommended that the operator
resets the Access Group of each employee.