Fixed bug #174
A new flag, _blockNextAction, is introduced to the cart controller.

Before entering the Deferred chain, this flag is set to true, and upon exiting
from the chain, the flag is cleared to false. The beforeFilter will check this
flag to decide if the currently requested operation ought to be blocked or not.