Fixed bug #206
Added new option 'vivipos.fec.settings.CartScrollMode', with (currently)
supported values of 'cursor-line' (the default) and 'view-page'.

In 'cursor-line' mode, scrolling the cart will move the cursor one line
in the indicated direction.

In 'view-page' mode, scrolling the cart will move the view one page in
the indicated direction, but the cursor will remain on the original
item.

In the future, 'cursor-page' and 'view-line' modes will also be
supported.