Fixed bug #340
When returning item from the cart, a 'returned' flag is set in the display
sequence entry for the returned item to differentiate the returned item from a
refund item. Each time an item is returned from the cart, we'll allow the
operation to proceed only if the available return quantity (by adding all
positive quantities and all negative quantities where the 'returned' flag is
set) is larger than or equal to the requested return amount.