Fixed bug #440
Bug caused by the fact that cashdrawer controller is not available in
the current scope. Added code to the Javascript startup files to invoke
the appropriate method on the cashdrawer controller in the main scope
for both the regular shift change screen and the shift change screen
requiring cash declaration.